Plot a bounded Voronoï 3D cell

Plot a bounded Voronoï 3D cell with rgl.

Usage

plotBoundedCell3D(
  cell,
  edgesAsTubes = FALSE,
  tubeRadius,
  tubeColor,
  facetsColor = NA,
  alpha = 1,
  check.bounded = TRUE
)

Arguments

cell

a bounded Voronoï 3D cell

edgesAsTubes

Boolean, whether to plot edges as tubes or as lines

tubeRadius

radius of the tubes if edgesAsTubes = TRUE

tubeColor

color of the tubes if edgesAsTubes = TRUE

facetsColor

color of the facets; NA for no color

alpha

opacity of the facets, a number between 0 and 1

check.bounded

Boolean, whether to check that the cell is bounded; set to FALSE for a small speed gain if you know that the cell is bounded

Value

No value, this function just plots the cell.

Examples

library(tessellation)
d <- delaunay(centricCuboctahedron())
v <- voronoi(d)
#> Voronoï diagram with one bounded cell.
cell13 <- v[[13]]
isBoundedCell(cell13) # TRUE
#> [1] TRUE
library(rgl)
open3d(windowRect = c(50, 50, 562, 562))
plotBoundedCell3D(
  cell13, edgesAsTubes = TRUE, tubeRadius = 0.03, tubeColor = "yellow",
  facetsColor = "navy", alpha = 0.7
)