Plot 3D Delaunay tessellation

Plot a 3D Delaunay tessellation with rgl.

Usage

plotDelaunay3D(
  tessellation,
  color = "distinct",
  hue = "random",
  luminosity = "light",
  alpha = 0.3,
  exteriorEdgesAsTubes = FALSE,
  tubeRadius,
  tubeColor
)

Arguments

tessellation

the output of delaunay

color

controls the filling colors of the tetrahedra, either FALSE for no color, "random" to use randomColor, or "distinct" to use distinctColorPalette

hue, luminosity

if color = "random", these arguments are passed to randomColor

alpha

opacity, number between 0 and 1

exteriorEdgesAsTubes

Boolean, whether to plot the exterior edges as tubes; in order to use this feature, you need to set exteriorEdges = TRUE in the delaunay function

tubeRadius

if exteriorEdgesAsTubes = TRUE, the radius of the tubes

tubeColor

if exteriorEdgesAsTubes = TRUE, the color of the tubes

Value

No value, just renders a 3D plot.

Examples

library(tessellation)
pts <- rbind(
  c(-5, -5,  16),
  c(-5,  8,   3),
  c(4,  -1,   3),
  c(4,  -5,   7),
  c(4,  -1, -10),
  c(4,  -5, -10),
  c(-5,  8, -10),
  c(-5, -5, -10)
)
tess <- delaunay(pts)
library(rgl)
open3d(windowRect = c(50, 50, 562, 562))
plotDelaunay3D(tess)
open3d(windowRect = c(50, 50, 562, 562))
plotDelaunay3D(
  tess, exteriorEdgesAsTubes = TRUE, tubeRadius = 0.3, tubeColor = "yellow"
)
#> Warning: You didn't set the option `exteriorEdges=TRUE` in the `delaunay` function, therefore the option `exteriorEdgesAsTubes` is ignored.