Delaunay triangulation

Delaunay triangulation (or tessellation) of a set of points.

Usage

delaunay(points, atinfinity = FALSE, degenerate = FALSE, exteriorEdges = FALSE)

Arguments

points

the points given as a matrix, one point per row

atinfinity

Boolean, whether to include a point at infinity

degenerate

Boolean, whether to include degenerate tiles

exteriorEdges

Boolean, for dimension 3 only, whether to return the exterior edges (see below)

Value

The Delaunay tessellation with many details, in a list. This list contains three fields:

vertices

the vertices (or sites) of the tessellation; these are the points passed to the function

tiles

the tiles of the tessellation (triangles in dimension 2, tetrahedra in dimension 3)

tilefacets

the facets of the tiles of the tessellation

In dimension 3, the list contains an additional field exteriorEdges if you set exteriorEdges = TRUE. This is the list of the exterior edges, represented as Edge3 objects. This field is involved in the function plotDelaunay3D. The vertices field is a list with the following fields:

id

the id of the vertex; this is nothing but the index of the corresponding point passed to the function

neighvertices

the ids of the vertices of the tessellation connected to this vertex by an edge

neightilefacets

the ids of the tile facets this vertex belongs to

neightiles

the ids of the tiles this vertex belongs to

The tiles field is a list with the following fields:

id

the id of the tile

simplex

a list describing the simplex (that is, the tile); this list contains four fields: vertices, a hash giving the simplex vertices and their id, circumcenter, the circumcenter of the simplex, circumradius, the circumradius of the simplex, and volume, the volume of the simplex

facets

the ids of the facets of this tile

neighbors

the ids of the tiles adjacent to this tile

family

two tiles have the same family if they share the same circumcenter; in this case the family is an integer, and the family is NA for tiles which do not share their circumcenter with any other tile

orientation

1 or -1, an indicator of the orientation of the tile

The tilefacets field is a list with the following fields:

id

the id of this tile facet

subsimplex

a list describing the subsimplex (that is, the tile facet); this list is similar to the simplex list of tiles

facetOf

one or two ids, the id(s) of the tile this facet belongs to

normal

a vector, the normal of the tile facet

offset

a number, the offset of the tile facet

See also

Examples

library(tessellation)
points <- rbind(
 c(0.5,0.5,0.5),
 c(0,0,0),
 c(0,0,1),
 c(0,1,0),
 c(0,1,1),
 c(1,0,0),
 c(1,0,1),
 c(1,1,0),
 c(1,1,1)
)
del <- delaunay(points)
del$vertices[[1]]
#> $id
#> [1] 1
#> 
#> $point
#> [1] 0.5 0.5 0.5
#> 
#> $neighvertices
#> [1] 2 3 4 5 6 7 8 9
#> 
#> $neightilefacets
#>  [1]  2  3  4  6  7  9 10 12 13 15 17 18 20 21 23 25 27 29
#> 
#> $neightiles
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12
#> 
del$tiles[[1]]
#> $id
#> [1] 1
#> 
#> $simplex
#> $simplex$vertices
#> <hash> containing 4 key-value pair(s).
#>   1 : 0.5 0.5 0.5
#>   2 : 0 0 0
#>   6 : 1 0 0
#>   8 : 1 1 0
#> 
#> $simplex$circumcenter
#> [1]  0.50  0.50 -0.25
#> 
#> $simplex$circumradius
#> [1] 0.75
#> 
#> $simplex$volume
#> [1] 0.08333333
#> 
#> 
#> $facets
#> [1] 1 2 3 4
#> 
#> $neighbors
#> [1] 3 2 5
#> 
#> $family
#> [1] 0
#> 
#> $orientation
#> [1] -1
#> 
del$tilefacets[[1]]
#> $id
#> [1] 1
#> 
#> $subsimplex
#> $subsimplex$vertices
#> <hash> containing 3 key-value pair(s).
#>   2 : 0 0 0
#>   6 : 1 0 0
#>   8 : 1 1 0
#> 
#> $subsimplex$circumcenter
#> [1] 0.5 0.5 0.0
#> 
#> $subsimplex$circumradius
#> [1] 0.7071068
#> 
#> $subsimplex$volume
#> [1] 0.5
#> 
#> 
#> $facetOf
#> [1] 1
#> 
#> $normal
#> [1] 0 0 1
#> 
#> $offset
#> [1] 0
#>